Refactor: Group different types of components into sub-dirs.
